+++ internal/provider/semvers_list_data_source.go
@@ -14,7 +14,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 
-	"github.com/anapsix/terraform-provider-semvers/internal/helpers"
+	shelper "github.com/anapsix/terraform-provider-semvers/internal/helpers"
 )
 
 // Data Source definition